split to files by hotkey type
